Fix a bug that was breaking the build in 32-bit mode.  For 32-bit, sizeof(int) = sizeof(long) and the build failed on "call of overloaded 'toStr(long int)' is ambiguous."
I'm not sure why StrUtil::toStr(long) is ambiguous, but it's easy
enough to fix.  Bad C++, bad, bad, awful, naughty C++.

+++ src/lib/banal/Struct.cpp
@@ -329,7 +329,7 @@ public:
   std::string
   toString() const
   {
-    std::string str = StrUtil::toStr((long)sortId()) + ":[" + filenm() + "]";
+    std::string str = StrUtil::toStr(sortId()) + ":[" + filenm() + "]";
     return str;
   }
